Update core and cvar

+++ src/toolkit/maths/hash/TkHash.h
@@ -6,12 +6,12 @@
 class cTkHash
 {
   public:
-    static uint64_t FNV1A(const void *lpData, uint64_t luiSize)
+    static u64 FNV1A(const void *lpData, u64 luiSize)
     {
-        uint64_t lHash = FNV_OFFSET;
-        for (size_t i = 0; i < luiSize; i++)
+        u64 lHash = FNV_OFFSET;
+        for (u64 i = 0; i < luiSize; i++)
         {
-            lHash ^= static_cast<const uint8_t *>(lpData)[i];
+            lHash ^= static_cast<const u8 *>(lpData)[i];
             lHash *= FNV_PRIME;
         }
         return lHash;
